Fix CodeQL high alerts: format-string mismatches + integer-mult casts (#203)
Real bugs:
- zserver_events.cpp x2: printf'd a size_t (player_info.size()) with %s -> %zu.
- gmm_player_list.cpp: sprintf'd a size_t with %d -> cast to int.

Overflow-safety (CodeQL cpp/integer-multiplication-cast-to-long): widen the
W*H / w*h products to size_t before they're used as container sizes / malloc
sizes / compared to .size(), in map_gen.cpp, zmap.cpp and zpath_finding.cpp.

The remaining alerts are by-design (plaintext game protocol; local server
config storing optional MySQL creds) or third-party (SDL_rotozoom) and are
dismissed on the scanning dashboard with reasons.

+++ src/gmm_player_list.cpp
@@ -58,7 +58,7 @@ void GMMPlayerList::SetupList()
 	}
 
 	//set player online text
-	sprintf(num_c, "%d", player_list.GetEntryList().size());
+	sprintf(num_c, "%d", (int)player_list.GetEntryList().size());
 	players_onnum_label.SetText(num_c);
 
 	//add the bots
